+++ examples/ip_rules/d-myip.tf
@@ -1,0 +1,8 @@
+data "http" "myip" {
+  url = "http://ip4.clara.net"
+}
+
+output "my_ip" {
+  description = "Current IPv4"
+  value       = "${data.http.myip.body}/32"
+}