+++ modules.tf
@@ -70,6 +70,7 @@ module "function_app" {
   authorized_ips                          = var.authorized_ips
   authorized_service_tags                 = var.authorized_service_tags
   authorized_subnet_ids                   = var.authorized_subnet_ids
+  ip_restriction_headers                  = var.ip_restriction_headers
   function_app_vnet_integration_enabled   = var.function_app_vnet_integration_enabled
   function_app_vnet_integration_subnet_id = var.function_app_vnet_integration_subnet_id