+++ CHANGELOG.md
@@ -1,8 +1,5 @@
 # Unreleased
 
-Breaking:
-  * AZ-595: SCM parameters now uses dedicated variables, like in `app-service-web` module
-
 Changed
   * AZ-572: Revamp examples and improve CI
   * AZ-592: Bump minimum AzureRM provider version to `v2.57`
@@ -10,6 +7,7 @@ Changed
 Added
   * AZ-588: VNet integration option
   * AZ-592: Support for the `site_config.ip_restrictions.headers` property
+  * AZ-595: SCM parameters now uses dedicated variables, like in `app-service-web` module
 
 Fixed
   * AZ-589: Avoid plan drift when specifying Diagnostic Settings categories